Move publishing to org.korge and update Kotlin dependencies to 2.3.20
- Toolchain was updated: Kotlin 2.3.20, AGP 8.13.1, Dokka 2.1.0, Gradle 8.14.4
- Custom Sonatype staging/release logic was replaced with
  com.vanniktech.maven.publish for signing + Central Portal publishing
- Namespace on maven central was changed to org.korge and Maven Central
  badge in README.md was updated
- GitHub Actions workflows were refreshed (Java 21, updated actions versions,
  new publish and dokka tasks, optional yarn lock upgrade)
- API check was updated
- Update korlibs dependencies:
    org.korge:korlibs-annotations:6.1.0
    org.korge:korlibs-audio-core:6.1.0
    org.korge:korlibs-concurrent:6.1.0
    org.korge:korlibs-ffi-legacy:6.1.1
    org.korge:korlibs-io:6.1.0
    org.korge:korlibs-io-vfs:6.1.0
    org.korge:korlibs-logger:6.1.0
    org.korge:korlibs-math-core:6.1.0
    org.korge:korlibs-math-vector:6.1.0
    org.korge:korlibs-time-core:6.1.0
    org.jetbrains.kotlinx:kotlinx-coroutines-test:1.10.2

+++ korlibs-audio/src/korlibs/audio/format/mod/MOD.kt
@@ -33,7 +33,7 @@ object MOD : BaseModuleTracker.Format("mod") {
     override fun createTracker(): BaseModuleTracker = Protracker()
     override suspend fun fastValidate(data: AsyncStream): Boolean {
         val bytes = data.sliceStart(1080).readBytesExact(4)
-        val signature = CharArray(4) { bytes[it].toChar() }.concatToString()
+        val signature = CharArray(4) { bytes[it].toInt().toChar() }.concatToString()
         val channels = when (signature) {
             "M.K.", "M!K!", "4CHN", "FLT4" -> 4
             "6CHN" -> 6
